refactor: `secure-parse-json.ts` -> `secure-json-parse.ts`
